more stragglers

+++ src/storage/recovery/abstract_log_provider.cpp
@@ -68,7 +68,7 @@ std::pair<LogRecord *, std::vector<byte *>> AbstractLogProvider::ReadNextRecord(
       }
 
       // Compute attr sizes
-      std::vector<uint8_t> attr_sizes;
+      std::vector<uint16_t> attr_sizes;
       attr_sizes.reserve(num_cols);
       for (uint16_t attr_idx = 0; attr_idx < num_cols; attr_idx++) {
         attr_sizes.push_back(StorageUtil::AttrSizeFromBoundaries(attr_size_boundaries, attr_idx));
